EightGodsBlade Added

+++ Code/Projectiles/StarBlast.cs
@@ -1,6 +1,9 @@
-using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
+using Terraria;
+using System;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 using UnbiddenMod.Code.Dusts;
 
 namespace UnbiddenMod.Code.Projectiles
@@ -20,11 +23,11 @@ namespace UnbiddenMod.Code.Projectiles
       projectile.aiStyle = 1;
       projectile.friendly = true;
       projectile.melee = true;
-      projectile.tileCollide = false;
+      projectile.tileCollide = true;
       projectile.ignoreWater = true;
       aiType = 0;
       projectile.timeLeft = 300;
-
+      projectile.penetrate = 3;
 
     }
 
@@ -40,11 +43,11 @@ namespace UnbiddenMod.Code.Projectiles
         projectile.soundDelay = 8;
         Main.PlaySound(SoundID.Item9, projectile.position);
       }
-      for (int i = 0 ; i < 200 ; i++)
+      for (int i = 0; i < 200; i++)
       {
         NPC target = Main.npc[i];
         //This will allow the projectile to only target hostile NPC's by referencing the variable, "target", above
-        if(target.active && !target.dontTakeDamage && !target.friendly)
+        if (target.active && !target.dontTakeDamage && !target.friendly)
         {
           //Finding the horizontal position of the target and adjusting trajectory accordingly
           float shootToX = target.position.X + (float)target.width * 0.5f - projectile.Center.X;
@@ -53,8 +56,8 @@ namespace UnbiddenMod.Code.Projectiles
           //  √ shootToX² + shootToY², using the Pythagorean Theorem to calculate the distance from the target
           float distance = (float)System.Math.Sqrt((double)(shootToX * shootToX + shootToY * shootToY));
 
-        //f, in this scenario, is a measurement of Pixel Distance
-        if (distance < 80f && !target.friendly && target.active)
+          //f, in this scenario, is a measurement of Pixel Distance
+          if (distance < 80f && !target.friendly && target.active)
           {
             distance = 3f / distance;
             shootToY *= distance * 5;
@@ -66,15 +69,37 @@ namespace UnbiddenMod.Code.Projectiles
         }
       }
     }
-    public override void OnHitNPC(NPC target, int damage, float knockback, bool crit) 
+    public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
     {
       Player player = Main.player[projectile.owner];
-      int healingAmount = damage / 30; //decrease the value 30 to increase heal, increase value to decrease. Or you can just replace damage/x with a set value to heal, instead of making it based on damage.
+      int healingAmount = damage / 60; //decrease the value 30 to increase heal, increase value to decrease. Or you can just replace damage/x with a set value to heal, instead of making it based on damage.
       player.statLife += healingAmount;
       player.HealEffect(healingAmount, true);
-      projectile.Kill();
+      projectile.penetrate--;
     }
 
+    public override bool OnTileCollide(Vector2 oldVelocity)
+    {
+      projectile.penetrate--;
+      if (projectile.penetrate <= 0)
+      {
+        projectile.Kill();
+      }
+      else
+      {
+        Collision.HitTiles(projectile.position + projectile.velocity, projectile.velocity, projectile.width, projectile.height);
+        Main.PlaySound(SoundID.Item10, projectile.position);
+        if (projectile.velocity.X != oldVelocity.X)
+        {
+          projectile.velocity.X = -oldVelocity.X;
+        }
+        if (projectile.velocity.Y != oldVelocity.Y)
+        {
+          projectile.velocity.Y = -oldVelocity.Y;
+        }
+      }
+      return false;
+    }
     public override void Kill(int timeLeft)
     {
